Update PixelResponse.cpp
Changed charge to a double

+++ source/PixelResponse.cpp
@@ -87,7 +87,7 @@ namespace Qpix
             std::vector<std::vector<int>> RESET_TRUTH_W;
 
             // seting up some parameters
-            int charge = 0;
+            double charge = 0;
             int pix_size = Pix_info[i].time.size();
             int pix_dex = 0;
             double current_time = 0;
@@ -117,7 +117,7 @@ namespace Qpix
                     while( current_time > pix_time )
                     {
                         trk_id_holder.push_back(Pix_info[i].Trk_ID[pix_dex]);
-                        charge += 1;
+                        charge += 1.0;
                         pix_dex += 1;
                         if (pix_dex >= pix_size){break; }
                         pix_time = Pix_info[i].time[pix_dex];
@@ -194,7 +194,7 @@ namespace Qpix
             std::vector<std::vector<int>> RESET_TRUTH_W;
 
             // seting up some parameters
-            int charge = 0;
+            double charge = 0;
             int pix_size = Pix_info[i].time.size();
             int pix_dex = 0;
             // int current_time = 0;
@@ -238,7 +238,7 @@ namespace Qpix
                     {
                                                              
                         trk_id_holder.push_back(Pix_info[i].Trk_ID[pix_dex]);
-                        charge += 1/Qpix_params->Sampling;
+                        charge += 1.0/Qpix_params->Sampling;
 
                         pix_dex += 1;
                         if (pix_dex >= pix_size){break; }
